チャットのメンバーカードをファイル分割 & アバターを表示
